Split createInteractions in multiple parts for enhanced readability
Relates: #80, #33
