+++ packages/plugins/Draw/src/types.ts
@@ -1,10 +1,11 @@
 import { Mutation, MutationTree } from 'vuex'
 import { FeatureCollection } from 'geojson'
-import VectorSource from 'ol/source/Vector'
+import { Feature } from 'ol'
 import Geometry from 'ol/geom/Geometry'
+import VectorLayer from 'ol/layer/Vector'
+import VectorSource from 'ol/source/Vector'
 import { StyleLike } from 'ol/style/Style'
 import { DrawConfiguration, DrawMode } from '@polar/lib-custom-types'
-import { Feature } from 'ol'
 
 // The options that can be given to an ol/VectorLayer. Somehow the direct import from ol doesn't work.
 // This is a copy with the things that we currently use
@@ -14,6 +15,11 @@ export interface PolarVectorOptions {
 }
 
 export type Mode = 'none' | 'draw' | 'edit' | 'delete'
+
+export interface CreateInteractionsPayload {
+  drawSource: VectorSource
+  drawLayer: VectorLayer<VectorSource>
+}
 
 export interface DrawState {
   mode: Mode
